Update save_blip.sh
Thankyou so much for this script. I was having a few issues and made updates as below. I'm now successfully downloading very fast.
My platform is Mac.

I have:
added a check for cover page existence (useful if the script is stopped and re-started later)
increased the javascript-delay to 10000
changed the success grep from 'complete' to Done
added a couple of informative date stamps and value of $result